Move has_condition to policy module
