+++ testsuite/tests/kuadrant/gateway/reconciliation/test_invalid_issuer_reference.py
@@ -4,7 +4,7 @@ import pytest
 
 from testsuite.gateway import CustomReference
 from testsuite.policy.tls_policy import TLSPolicy
-from testsuite.utils import has_condition
+from testsuite.policy import has_condition
 
 pytestmark = [pytest.mark.kuadrant_only, pytest.mark.tlspolicy]
